Move write_task_mapping_file to ddln.utils
